# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Secure User Authentication System API
  description: |
    ## Overview
    A production-grade JWT-based Authentication & Authorization Backend System.

    ### Features
    - JWT Access + Refresh Token rotation
    - Bcrypt password hashing (salt rounds: 12)
    - Account lockout after 5 failed attempts
    - Email verification & Password reset
    - Helmet.js security headers
    - Rate limiting (global + auth-specific)
    - Role-based access control (user / admin)

    ### Authentication
    Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    **Author:** Gudala Venkata Satyanarayana  
    **Assignment:** Secure User Authentication System  
    **Date:** May 2026
  version: 1.0.0
  contact:
    name: Gudala Venkata Satyanarayana
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
  - url: https://your-deployed-app.railway.app/api
    description: Production Server
...
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
    patch:
      tags: [Users]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
      responses:
        '200':
          description: Profile updated
    delete:
      tags: [Users]
      summary: Deactivate account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deactivated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...